+++ CHANGELOG.md
@@ -9,15 +9,18 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
-- FF-1429 - Updated FunFair.Test.Common to 1.8.2.400
-- FF-2652 - Update all the .NET components to .NET Core 3.1.302
-- FF-1429 - Updated TeamCity.VSTest.TestAdapter to 1.0.22
 ### Removed
 ### Deployment Changes
 
 <!--
 Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
 -->
+## [1.8.0] - 2020-07-20
+### Changed
+- FF-1429 - Updated FunFair.Test.Common to 1.8.2.400
+- FF-2652 - Update all the .NET components to .NET Core 3.1.302
+- FF-1429 - Updated TeamCity.VSTest.TestAdapter to 1.0.22
+
 ## [1.7.2] - 2019-07-08
 ### Added
 - FF-2616 - Check compiler errors in unit test source code